python/requirements.txt
requests
kafka-python
paramiko
PyYAML==5.4
fabrictestbed-extensions==1.2
python-chi